feat: renamed term-definition-lenght.tests.js file to terms.test.js and implemented casing test
